app/requirements.txt
flask
requests
nexmo
python-dotenv
